[VAS-800] Added unit tests and fixed test packages
