+++ src/test/java/it/gov/pagopa/apiconfig/ApplicationTest.java
@@ -1,8 +1,7 @@
-package it.gov.pagopa.apiconfig.selfcareintegration;
+package it.gov.pagopa.apiconfig;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import it.gov.pagopa.apiconfig.Application;
 import org.junit.jupiter.api.Test;
 
 class ApplicationTest {
