+++ src/test/java/it/gov/pagopa/apiconfig/selfcareintegration/controller/BrokerPspsControllerTest.java
@@ -1,6 +1,6 @@
-package it.gov.pagopa.apiconfig.controller;
+package it.gov.pagopa.apiconfig.selfcareintegration.controller;
 
-import static it.gov.pagopa.apiconfig.util.TestUtil.getMockChannelDetailsList;
+import static it.gov.pagopa.apiconfig.selfcareintegration.util.TestUtil.getMockChannelDetailsList;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
