+++ src/test/java/it/gov/pagopa/apiconfig/selfcareintegration/controller/BrokersControllerTest.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.apiconfig.controller;
+package it.gov.pagopa.apiconfig.selfcareintegration.controller;
 
 import it.gov.pagopa.apiconfig.Application;
 import it.gov.pagopa.apiconfig.selfcareintegration.service.BrokersService;
@@ -15,8 +15,8 @@ import org.springframework.test.web.servlet.MockMvc;
 
 import java.io.IOException;
 
-import static it.gov.pagopa.apiconfig.util.TestUtil.getMockCreditorInstitutionDetails;
-import static it.gov.pagopa.apiconfig.util.TestUtil.getMockStationDetailsList;
+import static it.gov.pagopa.apiconfig.selfcareintegration.util.TestUtil.getMockCreditorInstitutionDetails;
+import static it.gov.pagopa.apiconfig.selfcareintegration.util.TestUtil.getMockStationDetailsList;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
