+++ src/test/java/it/gov/pagopa/apiconfig/selfcareintegration/controller/PspControllerTest.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.apiconfig.controller;
+package it.gov.pagopa.apiconfig.selfcareintegration.controller;
 
 import it.gov.pagopa.apiconfig.Application;
 import it.gov.pagopa.apiconfig.selfcareintegration.service.PspService;
@@ -15,7 +15,7 @@ import org.springframework.test.web.servlet.MockMvc;
 
 import java.io.IOException;
 
-import static it.gov.pagopa.apiconfig.util.TestUtil.getMockChannelDetailsList;
+import static it.gov.pagopa.apiconfig.selfcareintegration.util.TestUtil.getMockChannelDetailsList;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
