+++ src/test/java/it/gov/pagopa/apiconfig/selfcareintegration/exception/ErrorHandlerTest.java
@@ -1,12 +1,10 @@
-package it.gov.pagopa.apiconfig.exception;
+package it.gov.pagopa.apiconfig.selfcareintegration.exception;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.mockito.Mockito.when;
 
 import it.gov.pagopa.apiconfig.Application;
-import it.gov.pagopa.apiconfig.selfcareintegration.exception.AppException;
-import it.gov.pagopa.apiconfig.selfcareintegration.exception.ErrorHandler;
 import it.gov.pagopa.apiconfig.selfcareintegration.model.ProblemJson;
 import java.sql.SQLException;
 import org.hibernate.exception.ConstraintViolationException;
