+++ src/test/java/it/gov/pagopa/apiconfig/selfcareintegration/service/HealthCheckServiceTest.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.apiconfig.service;
+package it.gov.pagopa.apiconfig.selfcareintegration.service;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -6,7 +6,6 @@ import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.when;
 
 import it.gov.pagopa.apiconfig.Application;
-import it.gov.pagopa.apiconfig.selfcareintegration.service.HealthCheckService;
 import it.gov.pagopa.apiconfig.starter.repository.HealthCheckRepository;
 import java.util.Optional;
 import org.junit.jupiter.api.Test;
